requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.16.3
openai==1.3.0
anthropic==0.7.0
groq==0.36.0
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0
tiktoken==0.5.1
numpy==1.24.3
pandas==2.1.3


